+++ CHANGELOG.md
@@ -44,7 +44,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Option to change the title (similar to PCB Variant)
 - Render_3D: Options to disable some technical layers and control the
   silkscreen clipping. (#282)
-- Internal BoM: Now you can aggregate components using CSV files. (See #248)
+- Internal BoM:
+  - Now you can aggregate components using CSV files. (See #248)
+  - Added some basic support for "Exclude from BoM" flag (See #316)
 - Now you can check PCB and schematic parity using the `update_xml` preflight
   (See #297)
 - New filters: